% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{check_boundary}
\alias{check_boundary}
\title{function to check the lat and long are within limits.}
\usage{
check_boundary(input, check_value, expected_greater = TRUE, warning_content)
}
\arguments{
\item{input}{the input lat or long value.}

\item{check_value}{the max or min value of lat or long.}

\item{expected_greater}{the input is expected to be greater or not.  Defalut:TRUE.}

\item{warning_content}{the content used to insert in the error message.}
}
\description{
function to check the lat and long are within limits.
}
