% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{check_list}
\alias{check_list}
\title{check the list in the dataset or not.}
\usage{
check_list(input, check_list_data, warning_content)
}
\arguments{
\item{input}{the list to check.}

\item{check_list_data}{the list in the datasets.}

\item{warning_content}{the content to be raised as an error.}
}
\description{
check the list in the dataset or not.
}
