% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{data_overview}
\alias{data_overview}
\title{Data overview for usfertilizer}
\usage{
data_overview(fertilizer_data)
}
\arguments{
\item{fertilizer_data}{the dataset were selected based on the filters.}
}
\description{
Provides a summary of filtered data.
}
\examples{
wake_fertilizer = get_data(counties = "wake")
data_overview(wake_fertilizer)


}
\keyword{fertilizer}
\keyword{summary}
