% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{get_data}
\alias{get_data}
\title{get_data}
\arguments{
\item{fert_type}{fertilizer type of interest, default: both Nitrogen and Phosphorus.}

\item{farm_type}{where the fertilizer was applied. e.g. farm, nonfarm.}

\item{years}{start year to show data, e.g. 1994, 2000:2005.}

\item{counties}{counties of interest, defalut: all avaible counties.}

\item{states}{states of interest, defalt: all avaialble states.}

\item{lat_max}{the maximum latitude of area of interest.}

\item{lat_min}{the minmum latitude of area of interest.}

\item{long_max}{the maximum longitude of area of interest.}

\item{long_min}{the minimum longitude of area of interest.}

\item{FIPS}{FIPS numbers of interest, defalut: all the counties.}

\item{overlap_state_county}{Logic. If true, the function will overlaping
the input of states and counties. If false, the function will return
results either in the states or in the counties.}

\item{combine_state_county}{Logic. If true, the county will be changed into
county, state, e.g. Wake, NC; If false, no changes.}

\item{remove_data}{Logic.  If true, the data will be removed for this fentch
and load it again.  For the sake of internet speed, the author set it
true.  Users can set it false to keep it in the RAM.}
}
\value{
A tibble with tidy data.
}
\description{
This function will return the full dataset (or a subset based on input parameters) of county-level estimates
 of fertilizer application quantity.  The fertilizer application type includes farm and non-farms.
 The data were estimated based on the sales of commecial fertilizer for each county originally from
 the Association of American Plant Food Control Officials (AAPFCO) commercial fertilizer sales data.
 Further details are available via: https://www.sciencebase.gov/catalog/item/5851b2d1e4b0f99207c4f238
}
\examples{
get_data(counties = "Wake")
get_data(years = 1990)
get_data(fert_type = "N", years = 2003, states = "NC", FIPSs = 37145)

}
\seealso{
\code{link(get_FIPS)}
}
\keyword{datasets}
\keyword{tidydata}
