% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FIPS.R
\name{get_FIPS}
\alias{get_FIPS}
\title{Get the FIPS code for the selected data.}
\usage{
get_FIPS(counties = NULL, states = NULL, overlap_state_county = TRUE,
  combine_state_county = TRUE)
}
\arguments{
\item{counties}{counties of interest, defalut: all avaible counties.}

\item{states}{states of interest, defalt: all avaialble states.}

\item{overlap_state_county}{Logic. If true, the function will overlaping
the input of states and counties. If false, the function will return
results either in the states or in the counties.}

\item{combine_state_county}{Logic. If true, the county will be changed into
county, state, e.g. Wake, NC; If false, no changes.}
}
\value{
A tibble with tidy data.
}
\description{
The function will fetch and return the FIPS for the counties of interest.
   As some counties will change due to various causes, it is easier to track the counties
   by FIPS code for a long-term analysis.  Description of FIPS codes is available via
   \code{[here](https://en.wikipedia.org/wiki/FIPS_county_code)}
}
\examples{
   get_FIPS(counties = "Wake", states = "NC")
   get_FIPS(states = "NC")

}
\seealso{
\code{link(get_data)}
}
