% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfertilizer.R
\docType{package}
\name{usfertilizer-package}
\alias{usfertilizer}
\alias{usfertilizer-package}
\title{usfertilizer: County-level estimates of fertilizer application in US}
\description{
This package, usfertilizer, is used to explore and visualize the
county-level estimates of fertilizer, nitrogen and phosphorus, from 1987
to 2012.  The dataset were originally generated by USGS based on the sales
data of commecial fertilizer. See \code{
[here](https://www.sciencebase.gov/catalog/item/5851b2d1e4b0f99207c4f238/) }for details.
Usfertilizer utilized the tidyverse to clean the original data and provide
user-friendly functitons to select the data of interest, summarize dataset
and gain insights.  Please note that USGS does not endorse this package.
}
\keyword{internal}
