% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usfertilizer.R
\docType{package}
\name{usfertilizer-package}
\alias{usfertilizer}
\alias{usfertilizer-package}
\title{usfertilizer: County-Level Estimates of Fertilizer Application in US}
\description{
Compiled and cleaned the county-level estimates of fertilizer,
nitrogen and phosphorus, from 1945 to 2012.  The dataset were originally
generated by USGS based on the sales data of commercial fertilizer.
See <https://www.sciencebase.gov/catalog/item/5851b2d1e4b0f99207c4f238/> for details.
'usfertilizer' utilized the tidyverse to clean the original data and provide
user-friendly dataframe. Please note that USGS does not endorse this package and this package
does not include manure data. Also data from 1986 is not available for now.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wenlong-liu/usfertilizer}
  \item Report bugs at \url{https://github.com/wenlong-liu/usfertilizer/issues}
}

}
\author{
\strong{Maintainer}: Wenlong Liu \email{wliu14@ncsu.edu}

}
\keyword{internal}
